rafactor: ファイルの構成を変更
