+++ PrefeFortune/Views/FortuneResult/PrefectureImage/FailureView.swift
