+++ PrefeFortune/Views/FortuneResult/PrefectureImage/PlaceholderView.swift
