+++ PrefeFortune/Views/FortuneResult/TouristCard/PlaceImageView.swift
