+++ PrefeFortune/Views/FortuneResult/TouristCard/TouristCard.swift
