+++ PrefeFortune/Views/FortuneResult/TouristCard/TouristCardView.swift
